+++ CHANGELOG.md
@@ -3,9 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ## 0.4.0.0 (in development)
 - [#57] FIXED: Comment in code block breaks header detection for Haskell/Haddock
-- [#58] (BREAKING CHANGE) Extend configuration for defining header margins
+- [#58] BREAKING CHANGE: Extend configuration for defining header margins
 - [#59] Check compatibility of loaded YAML configuration
-- Bump _LTS Haskell_ to `16.25`
+- Bump _LTS Haskell_ to `16.26`
 
 ## 0.3.2.0 (released 2020-11-26)
 - [#56] Move CI from Travis to GitHub Actions