+++ CHANGELOG.md
@@ -5,8 +5,9 @@ All notable changes to this project will be documented in this file.
 - [#57] FIXED: Comment in code block breaks header detection for Haskell/Haddock
 - [#58] BREAKING CHANGE: Extend configuration for defining header margins
 - [#59] Check compatibility of loaded YAML configuration
+- [#61] BREAKING CHANGE: Use regular expressions to specify headers start/end
 - [#62] Sanitize block comment headers with line prefixes
-- Bump _LTS Haskell_ to `16.26`
+- Bump _LTS Haskell_ to `16.27`
 
 ## 0.3.2.0 (released 2020-11-26)
 - [#56] Move CI from Travis to GitHub Actions