+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 - [#59] Check compatibility of loaded YAML configuration
 - [#61] BREAKING CHANGE: Use regular expressions to specify headers start/end
 - [#62] Sanitize block comment headers with line prefixes
-- Bump _LTS Haskell_ to `17.1`
+- Bump _LTS Haskell_ to `17.2`
 
 ## 0.3.2.0 (released 2020-11-26)
 - [#56] Move CI from Travis to GitHub Actions