+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file.
 - [#61] BREAKING CHANGE: Use regular expressions to specify headers start/end
 - [#62] Sanitize block comment headers with line prefixes
 - [#64] Add command line option for displaying version
-- Bump _LTS Haskell_ to `17.4`
+- Bump _LTS Haskell_ to `17.5`
 
 ## 0.3.2.0 (released 2020-11-26)
 - [#56] Move CI from Travis to GitHub Actions