+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## 0.4.0.0 (in development)
+## 0.4.1.0 (upcoming release)
+
+## 0.4.0.0 (released 2021-03-04)
 - [#57] FIXED: put-before/put-after in commented code breaks header detection
 - [#58] BREAKING CHANGE: Extend configuration for defining header margins
 - [#59] Check compatibility of loaded YAML configuration